pyproject.toml
[project]
name = "fashion-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
